django-api/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
PyJWT==2.1.0
django-cors-headers==3.7.0
cloudinary==1.37.0
django-filter==2.4.0
django-fsm==2.8.0